requirements.txt
# Cerebras Deep Search Dependencies

# Core dependencies
exa-py>=1.14.0
cerebras-cloud-sdk>=1.35.0
python-dotenv>=1.0.0

# Development and testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
